+++ tests/builds/tb/Snakefile
@@ -206,7 +206,7 @@ rule export:
             --output-tree {output.tree} \
             --output-meta {output.meta} \
             --output-sequence {output.seq}
-        augur validate --json {output.tree} {output.meta}
+        augur validate v1 --json {output.tree} {output.meta}
         """
 
 
@@ -249,7 +249,7 @@ rule exportv2:
             --maintainers {params.auth} \
             --maintainer-urls {params.url} \
             --geography-traits {params.geo}
-        augur validate --new-schema --json {output.main}
+        augur validate v2 --json {output.main}
         """
 
 rule clean:
